examples/openai-responses/requirements.txt
openai
httpx
web3
python-dotenv
